Started to implement file specific settings
- Added support for adding project-specific settings files.  This is a
catch-all term that refers to solution, project, folder, and
file-specific settings files.  Files are added using the File | New |
Spell Checker Configuration File for Selected Item or by right clicking
an item in the Solution Explorer and selection the Spell Checker
Configuration File option on the main context menu or the Add sub-menu
(based on the item clicked).
- Converted the configuration dialog into an editor window.  This allows
it to edit the global configuration as well as the project-specific
settings files.  When editing file related settings, properties can be
indicated as inherited.
- NOTE: Support for using the file related settings has not yet been
implemented.  That will be done next.